Changes for getMyAgents and getMySupervisors
