+++ controllers/account-setting/my-team/getMyAgents.controller.js
@@ -1,25 +1,15 @@
-const MyTeamModel = require('../../../models/my-team.model');
+const { getAgentsListForMyTeam } = require('../../../services/myteam.service');
 
 const getMyAgentsController = async (req, res, next) => {
 	try {
 		const { user } = req.params;
 
-		const myTeams = await MyTeamModel.findOne({ user }).populate({
-			path: 'team.user'
-		});
-
-		if (myTeams) {
-			const supervisorList = myTeams.team.map((item) => {
-				//Until implement user roles and status
-				if (item.user.role === 'user') {
-					return { _id: item.user._id, fullname: item.user.fullname, email: item.user.email };
-				}
-			});
-
+		const data = await getAgentsListForMyTeam(user);
+		if (data.length) {
 			return res.status(200).send({
 				code: res.statusCode,
 				message: 'Agents found',
-				agents: supervisorList
+				agents: data
 			});
 		} else {
 			return res.status(404).send({
