+++ controllers/account-setting/my-team/getMySupervisors.controller.js
@@ -1,33 +1,26 @@
 const MyTeamModel = require('../../../models/my-team.model');
+const { getSupervisorsListForMyTeam } = require('../../../services/myteam.service');
 
 const getMySupervisorsController = async (req, res, next) => {
 	try {
 		const { user } = req.params;
 
-		const myTeams = await MyTeamModel.findOne({ user }).populate({
-			path: 'team.user'
-		});
-
-		if (myTeams) {
-			const supervisorList = myTeams.team.map((item) => {
-				//Until implement user roles and status
-				if (item.user.role === 'user') {
-					return { _id: item.user._id, fullname: item.user.fullname, email: item.user.email };
-				}
-			});
-
+		const data = await getSupervisorsListForMyTeam(user)
+		if(data.length){
 			return res.status(200).send({
 				code: res.statusCode,
 				message: 'Supervisors found',
-				supervisors: supervisorList
+				supervisors: data
 			});
-		} else {
+		}else{
 			return res.status(404).send({
 				code: res.statusCode,
 				message: 'Supervisors not found',
 				supervisors: []
 			});
 		}
+
+
 	} catch (error) {
 		return res.status(500).send({
 			code: 500,
